3eme

+++ ft_strncat.c
@@ -6,13 +6,13 @@
 /*   By: fpolini <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/11/27 17:56:49 by fpolini           #+#    #+#             */
-/*   Updated: 2015/11/27 18:33:46 by fpolini          ###   ########.fr       */
+/*   Updated: 2015/11/29 13:45:26 by fpolini          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_strncat(char *s1, char *s2, unsigned long n)
+char	*ft_strncat(char *s1, char *s2, size_t n)
 {
 	int	i;
 	int	j;
